Automatically create toolchain update PRs
This workflow will attempt to move the toolchain nightly pointer by one
day every four hours. In doing so it will create a pull request for just
this change. Once that PR is merged, further runs of this workflow will
then try to move the pointer even further. (And, thus, running the
workflow multiple times a day gives us a chance to catch up.)